requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.5
django-address==0.2.5
django-allauth==0.44.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangocms-admin-style==2.0.2
Flask==1.0.2
Flask-Compress==1.4.0
Flask-Gravatar==0.5.0
Flask-Login==0.4.1
Flask-Migrate==2.4.0
Flask-Paranoid==0.2.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.1.0
idna==2.10
MouseInfo==0.1.3
oauthlib==3.1.0
olefile==0.46
pgadmin4==4.29
Pillow==8.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.6
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
PyJWT==2.0.1
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python3-openid==3.2.0
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.5
pywhatkit==3.9
requests==2.25.1
requests-oauthlib==1.3.0
simplejson==3.16.0
soupsieve==2.2
speaklater==1.3
sqlparse==0.4.1
sshtunnel==0.4.0
urllib3==1.26.3
whitenoise==5.2.0
wikipedia==1.4.0
